pyproject.toml
[tool.poetry]
name = "mapp_nocodb_connector"
version = "0.1.0"
description = "A Python script to echange with the MAPP NocoDB instance"
authors = ["PMA <pierre-marie.allard@unifr.ch>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
argparse = "^1.4.0"
pandas = "^2.1.4"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
